package/alsa-lib/alsa-lib.mk package/gettext/gettext.mk: Both of these components look for iconv when configuring so need to be dependent.
Both of these packages can be built with libiconv support.  If libiconv is enabled and these are built first 
then some configure variables get stored in the system cache and then libiconv is not used properly.
If libiconv is selected by user make sure it is built before either of these packages are built.

Signed-off-by: Daniel Laird <daniel.j.laird@nxp.com>