package/libidn: Add a new package (required by later WebKits...)
GNU Libidn is a fully documented implementation of the Stringprep, 
Punycode and IDNA specifications. Libidn's purpose is to encode 
and decode internationalized domain names. The native C, C# 
and Java libraries are available under 
the GNU Lesser General Public License version 2.1 or later. 

Not quite sure I've put it in the correct menu but it will work.

Signed-off-by: Daniel Laird <daniel.j.laird@nxp.com>