documentation: insist more on why output/target/ should not be used
A lot of people complain that the root filesystem generated by
Buildroot doesn't work because it lacks the device files. This is
because they use output/target/ instead of one of the images.

Therefore, insist in the documentation on this point. Even though I'm
unsure if users really read the documentation :-)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>