Remove all TARGET_DEVICE_TABLE definitions
We have a special case for Xtensa, which was patching the generic
device_table.txt. Instead of doing this, we just keep a copy of the
device table, specific to Xtensa, with Xtensa specifities. The fact
that the patch wasn't applying anymore on the generic device table is
a sign that the existing approach wasn't working anyway.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>