xfonts: fix dependencies and build process
Instead of assuming that mkfontdir and mkfontscale are available on
the development system, use the one installed in $(HOST_DIR).

The FCCACHE variable is removed, because, for some reason, the
installation process of xfonts disables the call to fc-cache when a
DESTDIR is passed. See the definition of XORG_FONT_FCCACHE in
aclocal.m4 in font packages:

AC_DEFUN([XORG_FONT_FCCACHE],[
        AC_PATH_PROG(FCCACHE, fc-cache)
        FCCACHE_WARN='echo "** Warning: fonts.cache not built" ; echo "** Generate this file manually on host system using fc-cache"'
        if test x"$FCCACHE" = x ; then
                RUN_FCCACHE="${FCCACHE_WARN}"
        else
                RUN_FCCACHE='@(if test -z "$(DESTDIR)"; then echo $(FCCACHE) $(fontdir); $(FCCACHE) $(fontdir); else'
                RUN_FCCACHE="${RUN_FCCACHE} ${FCCACHE_WARN} ; fi)"
        fi
        AC_SUBST([RUN_FCCACHE])
])

In addition to this, we update the font dependencies to include
font-utils for the host, mkfontdir for the host, mkfontscale for the
host and bdftopcf for the host.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>