xkeyboard-config: add dependency on host-xapp_xkbcomp
In order to get xkeyboard-config to build, it needs xkbcomp to be
built on the host, i.e host-xapp_xkbcomp. In turn, xapp_xkbcomp needs
xlib_libX11 and xlib_libxkbfile, which in turn need a few libraries.

This commit adds all the relevant dependencies to build everything
that is needed to get xkeyboard-config to build correctly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>