u-boot: fix custom patch handling
When U-Boot is enabled and no custom patch directory has been set,
then the current test:

  ifneq ($(strip $(BR2_TARGET_UBOOT_CUSTOM_PATCH_DIR)),"")

works. However, when U-Boot is not enabled, but still gets compiled
because mkimage is needed to build the kernel,
BR2_TARGET_UBOOT_CUSTOM_PATCH_DIR is completely empty. It does not
even have quotes. So the test in fact needs to be:

 ifneq ($(call qstrip,$(BR2_TARGET_UBOOT_CUSTOM_PATCH_DIR)),)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>