libglib2: remove fix-clock-gettime patch
This patch was introduced in 0ec4298153bd93f8d7efbd1890896fbac5046a7b
to fix a problem in libglib configure script when using the
configuration cache.

However, we decided that the shared configuration cache cannot work
reliably, to disable it by default, and probably to remove it
altogether. So just drop this patch, and don't autoreconfigure
libglib.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>