lsof: stop overriding CFLAGS in build
Closes #2767

lsof compile breaks if CFLAGS overridden in 'make' command line
which is the case since buildroot-2010.08.

lsof has a partial fix for this already in the form of the
'LSOF_CFLAGS_OVERRIDE=1' configure option; however this only
fixes the problem in the generation of the Makefile for liblsof.

This patch does two things: it extends the 'LSOF_CFLAGS_OVERRIDE=1'
configure option to also cover the main Makefile; it also patches
lsof.mk to make use of the LSOF_CFLAGS_OVERRIDE option.

Signed-off-by: Andy Gibbs <andyg1001@hotmail.co.uk>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>