board: remove support for at91sam9260pf
It was supposed to be the support for AT91SAM9260 using a parallel flash
(instead of the usual dataflash). But the provided U-Boot configuration
at91sam9260pf.h was not used anywhere, and it was unsufficient to add
correct support in U-Boot (some changes in U-Boot Makefile would also be
needed). Additionnally, this configuration has not been merged into U-Boot
upstream since 2007 (when it was added to Buildroot).

Therefore, let's get rid of this configuration. If some users are
interested, we can re-introduce it properly with their help.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>