linux: don't force IPv6 configuration in kernel
As per the discussion at
http://lists.busybox.net/pipermail/buildroot/2010-December/040030.html,
remove the bits that forced the IPv6 configuration in the kernel
depending on the toolchain ability to support (or not) IPv6. You may
have a toolchain with IPv6 support but still don't want to have IPv6
in your kernel.

The only parameters we adjust in the kernel configuration are:

 * ARM EABI, since we got a lot of bug reports regarding misconfigured
   kernel compared to the ABI used by userspace applications.

 * initramfs, since its contents are generated by Buildroot itself

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>