x11r7: make package globally visible
All x11r7 .mk files were only included into the global chain of
makefile includes if the BR2_PACKAGE_XORG7 option was enabled. This is
not the standard way to do things in Buildroot, and it even breaks
things like running "make <pkg>-*" for a x11r7 package even if X.org
is not selected.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>