netkittelnet: fix build on toolchains without C++ support
The configure script checks for C++ support as telnet (but not telnetd)
contains C++ code. This used to work by accident if the host had g++,
as we didn't pass TARGET_CONFIGURE_OPTS to configure (which sets CXX to
/bin/false).

Fix it by removing the g++ check and rewriting the configure checks to
use the C compiler instead.

Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>