strace: fix build for x86_64
strace fails to build on x86_64 because stat64 is not available.  This
is because the automatic detection of stat64 in configure is overridden
by buildroot, by setting ac_cv_type_stat64.  Just remove that override -
current strace seems to detect it correctly for non-largefile platforms.

Build-tested on x86_64 (with largefile), ARM (with and without largefile),
sh4, MIPS and ppc-32 (no largefile).

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>