librsvg: relax dependency on Gtk, add gdk-pixbuf dependency
librsvg doesn't *require* Gtk, it can optionally build a Gtk theme
engine, but it is not mandatory. Therefore, we make the Gtk dependency
an optional dependency rather than a hard dependency. This will be
useful as librsvg will become a dependency of libevas-generic-loaders,
used in an EFL context in which building libgtk is not really what we
want.

However, gdk-pixbuf is a mandatory dependency to build librsvg, so we
add this one.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>