xserver_xorg-server: rename server-style configuration options
The BR2_PACKAGE_XSERVER_xorg and BR2_PACKAGE_XSERVER_tinyx options
used to select the style of X.org server to use are not named
consistently with the rest of the Buildroot options (in capital
letters and prefixed with the package name).

Therefore, we rename those options, and we take care to add the old
option names in the BR2_LEGACY infrastructure.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Acked-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>