atk: remove unused host variant
The atk package is used as a dependency in the libglade and libgtk2
packages.

The libglade package has no host variant, and does not depend on
host-atk.

The libgtk2 package depends on atk, and has a host variant, but the
host variant of libgtk2 overrides HOST_LIBGTK2_DEPENDENCIES to limit
the number of dependencies, and host-atk is not amongst those
dependencies.

Therefore, host-atk is useless and can be removed.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>