rpi-userland: update to fix no-longer applying patch
Our patches do not apply as-is against the version of rpi-userland we
package (weird...).

Since rpi-userland does not have releases, and is a relatively fast-moving
target, just bump the version to the latest cset from the git tree, and
get rid of one our failing patch, since it was applied upstream.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Maxime Hadjinlian <maxime.hadjinlian@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>