gptfdisk: ensure that the package always builds/installs something
Until now, selecting just the gptfdisk package wouldn't do anything.
It wouldn't build anything at all, and wouldn't install anything.
This is rather odd.

So now, we ensure that at least one of the two tools that gptfdisk can
build and install is selected. This allows use to factorize a bit the
select of util-linux in Config.in, and remove the if condition in the
makefile that was preventing the package from being built if no tools
was selected.

[Peter: Reword commit text]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>