directfb: fix error in *SetRegion* datatypes
DirectFB comes with incorrect declarations for *SetRegion* datatypes,
this breaks the build for some GFX Drivers. Also there are some headers
missend in the archive, this breaks some other modules of directfb that
not covered yet by the autobuilders.
And at least the configure script doesn't use a variable for the
imlib2-config script. That breaks crossbuilds of directfb in most cases.

Fixes:
http://autobuild.buildroot.net/results/03465f0e14accc8d8f8fc2640b7a0dd8bec594e4

Signed-off-by: Carsten Schoenert <c.schoenert@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>