gutenprint: ensure target and host variant have the same patch hooks
In preparation for the out-of-tree support, it is important that all
packages have the same patch hooks for the host variant and the target
variant (because the source tree will be shared between host and
target variants when doing out-of-tree build).

In the case of gutenprint, the patch hook executed for the target just
creates one directory which is needed for autoreconf to work, and the
host variant is not autoreconf'ed. So there is no harm in also
applying this patch hook to the host variant.

[Peter: autoreconf for host as pointed out by Arnout]
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>