p11-kit: fix argp_standalone build error
When BR2_PACKAGE_ARGP_STANDALONE is enabled and built before p11-kit the
configure script detects it and a conflict arises because libc already
includes and defines program_invocation_short_name.
Avoid this by forcing the detection to true otherwise the configure
script tries to use argp-standalone and it clashes with libc. Fixes:
http://autobuild.buildroot.net/results/a39/a39fe81a356b922a891d4225a56ce2427ab5bca7/

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>