lttng: bump the various LTTng components
This commit bumps the various LTTng components to their latest stable
version: 0.7.7 for liburcu, 1.1.1 for babeltrace, 2.1.3 for libust and
2.1.2 for lttng-modules.

For babeltrace, the patch
lttng-babeltrace-02-fix-build-old-compilers.patch has been removed
because it has been merged upstream as of version 1.1.1. This is also
the reason why the package no longer needs to be autoreconfigured.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>