legacy: add a legacy option for BR2_ARM_OABI
As Arnout noted, we removed the BR2_ARM_OABI option without adding the
corresponding option in Config.in.legacy, so users upgrading to
Buildroot 2013.08 with an OABI configuration would get migrated
automatically to EABI without any warning.

This commit introduces such a legacy option, so that such users would
be explicitly warned about the removal of OABI support.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>