trivial: correct capitalization of Lua in Config.in files
The Lua website says:
"Like most names, it should be written in lower case with an initial capital,
that is, "Lua". Please do not write it as "LUA", which is both ugly and
confusing, because then it becomes an acronym with different meanings for
different people."
http://www.lua.org/about.html

So, let's honor this request in buildroot.

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Signed-off-by: Peter Korsgaard <jacmet@sunsite.dk>