omniorb: cleanup autobuild failure, CFLAGS issue
Fixed:
http://autobuild.buildroot.net/results/a08baf8821e3ee1fb9c36c8f74c4ced466af167b/
http://autobuild.buildroot.net/results/f0c7379c1f405385d3a64c0349c3d9565065e8a7/
http://autobuild.buildroot.net/results/c3de25ca52fdb8ffda7472cf9f7b64514954cbb1/

Change Log:
The target CFLAGS were getting mixed in during a manual build of some
omniorb pkg tools that could be viewed as host tools.
This manual build was a dirty workaround and created some bad
cornercases. So I converted the package over to actually using the
host pkg build and removed the need for a patch and custom mods in the build.

Also there wasn't a dependency on the target having python, just
host-python for performing the initial build.  So depends were adjusted
accordingly. If python bindings are required, they can be found in a different
package called omniORBpy.  The omniORB package only uses python for
the build process.

Fixed a bug with long double support.  Currently leaving it disabled
as most targets won't use it.  This allows greater toolchain compatibility.

Signed-off-by: Matt Weber <mlweber1@rockwellcollins.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>