perf: disable documentation
perf uses asciidoc to generate documentation, but /usr/bin/asciidoc will
try to use buildroot's host-python which doesn't have the required
unicodedata. Recent kernels always try to build the man pages when
installing. To avoid that, define ASCIIDOC as an empty string so the
Makefile can't find the executable and skips the documentation
generation.

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>