openpgm: disable for AVR32, not its Buildroot-built toolchain
OpenPGM builds incorrectly on AVR32 with gcc-4.2.2-avr32-2.1.5.
Since it is presumed to be the only GCC version used by Buildroot
for AVR32, this patch disables openpgm for all cases when AVR32
is selected as the target architecture, including when a toolchain
is downloaded or preinstalled (this is what Buildroot autobuilders
do).

Signed-off-by: Alexander Lukichev <alexander.lukichev@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>