gpsd: do not allow profiling support on AArch64
The profiling support is not available on AArch64, causing build
failures such as:

  http://autobuild.buildroot.net/results/724f8646e1c2c28cf187f2313244d1e3e61b4c25/

According to Will Newton, function profiling will only be added in the
next version of the AArch64 Linaro toolchain. This indicates it is a
rather new feature in AArch64, so for the moment, just disallow
profiling on AArch64. We can revise this later, when this feature has
been made available in Linaro toolchains, and official gcc releases
(for internal toolchain support).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>