package/linux-firmware: install license files with 'make legal-info'
Currently, the license files for the many firmware files are installed
in the target, alongside the firmware files, and are not available to the
legal-info infrastructture.

This patch separates the license files from the firmware files, and makes
them available to the legal-info infrastructure, so they get installed
with all the other license files.

Since most firmware files are proprietary (at least those we currently
install), we just have a single license definition. When we later support
other firmwares that have an OSS license, those can add their own
licensing information on a case-by-case basis.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>