pcsc-lite: add patch to fix libusb detection when linking statically
When libusb support is requested in pcsc-lite, the configure script will try
to link a test application to check libusb_init(). However, when libusb
needs threads, the linking step should link in threads as well.

The configure script, first tries to call libusb-config to determine the
necessary cflags and libraries (which should include the thread flags), but
this program does not exist, so the cflags/libs variables are empty.

Therefore, add a patch to explicitly add PTHREAD_CFLAGS and PTHREAD_LIBS to
the libusb test, so libusb is detected correctly.

Fixes
http://autobuild.buildroot.org/results/432/432be0db3f8415f6a026b69234579ad4d23b6421/

Signed-off-by: Thomas De Schampheleire <thomas.de.schampheleire@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>