libtool: use a different way to avoid autoreconfiguration
In 4268d39, a patch was added to update libtool.m4 for MIPS n64 targets.
However, this caused the configure script to be regenerated during the
build steps, which in turn caused build failures on Fedora 12 hosts. In
ed73d1d, this was fixed by patching the installed libtool.m4 file
instead of the source file.

However, ed73d1d did not take into account that the target libtool will
also install libtool.m4 to staging, and that that version has
precedence over the one in HOST_DIR.

Therefore, this patch takes a different approach: the source file is
patched in the usual way, and instead the reconfigure is avoided by
touching the generated files.

On the target, we also cannot use the AUTORECONF mechanism because of
some mysterious Makefile.inc file. So just use the same approach and
wait for upstream to make a new release.

Fixes:
   http://autobuild.buildroot.net/results/b86/b86a83c6549004f226e7255242e54ef4e50c5ec3/

Signed-off-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>