qemu x86/x86_64: Add kernel IPC support
The alsa utils packages require a kernel with IPC support to run.  The
kernel config for the x86 and x86_64 qemu build includes sound support
using the hda driver but without the IPC support the alsa utils won't work.
This patch adds IPC support to the x86 and x86_64 kernel to allow alsa
utils to run and brings the kernel config inline with kernel configs for
other qemu boards.

Signed-off-by: Martin Bark <martin@barkynet.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>