ipsec-tools: needs MMU support
ipsec-tools currently fails to build for Blackfin due to GLOB_TILDE
not being supported in Blackfin external toolchains. However, after
fixing this problem (which will be fixed in later versions of the
Blackfin external toolchains) allows to discover that ipsec-tools uses
fork() anyway.

Therefore, this patch simply marks ipsec-tools as not available on
non-MMU systems.

Consequently, it fixes:

  http://autobuild.buildroot.org/results/fc4/fc4ef4101a20e64eb52da56557d6996dd73b3d86/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>