package/trace-cmd: bump to version 2.3.2
The first patch "use-pkg-config-instead-of-python-config" has been applied
upstream.

The flag _GNU_SOURCE is now always defined in Makefile, so it can be removed
from trace-cmd.mk

CPPFLAGS are appended to CFLAGS and are used to add extra flags, but the flag
_LARGEFILE64_SOURCE in TARGET_CPPFLAGS is already defined in source files,
which causes a build error. As for CFLAGS, we fix this by filtering out our
definition of _LARGEFILE64_SOURCE from the CPPFLAGS before passing them to the
trace-cmd Makefile.

Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>