python-flup: re-add MIT license
The MIT license was in the original submission, but I had removed it
since I couldn't find where it was used. It is indeed used in
flup/server/paste_factory.py as noted by the original
submitter. Therefore, this commit adds the MIT license as one of the
licenses of python-flup.

Reported-by: Oli Vogt <oli.vogt.pub01@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>