perl-gd: mark as broken
perl-gd doesn't build on gcc10 due to the host Perl being a bit
old. This shows a defiency in the perl package infrastructure, which
is being worked on. However, since it won't be solved before 2014.08,
our only solution for now is to mark perl-gd as broken, as well as its
reverse dependency perl-gdgraph.

Avoids:

  http://autobuild.buildroot.org/results/228/228b43e67e683cecfa3851a8a030d06e9fdc6dac/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>