package/mesa3d: fix comment about i965 DRI driver
The help entry for the i965 DRI driver states that it supports alll
Intel GPUs. That's not true, as at least the i915 series is not handled
by this driver, but by a separate one.

Fix the help text accordingly; remove the useless second part of the
help text (which does not mean much for Buildroot.)

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>