ffmpeg: consolidate duplicated config options
Commit 62ab07ef769bd6504fe1db144aaac3fd45db9dad added a bunch
of hard-coded options.  Some of these options are then again
added (or negated) based on kconfig settings.  This patch removes
the hard coded swscale and postproc settings, as they are enabled
or disabled later.  It also moves the --disable-pic lower
as an else clause to the part that adds --enable-pic.

Signed-off-by: Danomi Manchego <danomimanchego123@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>