package/pkg-autotools.mk: set --localstatedir=/var by default
`localstatedir` is a common autoconf variable like prefix and sysconfdir. At
this time 25 autotools and one generic package uses this Makefile variable. 22
of the autotools packages use --localstatedir=/var, the remaining use custom
settings.

Setting `localstatedir` to /var is consistent with the Filesystem Hierarchy
Standard (FHS). So it's time to move it as a common definition in the
autotools infrastructure.

Signed-off-by: Jörg Krause <jkrause@posteo.de>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>