xtables-addons: new package
The geoip "helpers" are basically scripts that download and reformat
the geoip database in a form usable by xt_geoip.

The netfilter (kernel & userland) sides of it are built and installed.
Since there are many considerations to geoip databases (free,
commercial and variants for each) it's left to the user to deal with
that if they plan to use the extension which is only one among many.

[Thomas:
 - Take into account the rename of BR2_PREFER_STATIC_LIB to
   BR2_STATIC_LIBS
 - Remove "depends on BR2_LINUX_KERNEL" as suggested by Arnout.
 - Move XTABLES_ADDONS_CONF_OPTS a bit further down, with newlines
   around it, and adjust the indentation of the first line. Just to
   make it slightly more readable.]

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>