package/rpm: add missing dependency to openssl
We forcibly enable use of an extrernal openssl in rpm, so we must depend
on it.

openssl is correctly selected from kconfig, but there is no build-time
dependency to it.

It was not an issue so far, because openssl always sorts before rpm, so
gets built before. Unless one wants to just build rpm to debug an
autobuilder failure, that is.

Add that build-time dependency to openssl.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>