dhcpcd: use standard installation
Drop the custom/manual install cmds and do a normal make install.
This gets dhcpcd-hooks installed which are useful for many accesory
setup functions like dns, ntp server, and so on.
Also install dhcpcd to /sbin since it's expected by debian ifupdown,
For busybox ifupdown when external dhcp clients are enabled (not in the
default config we ship) will search the path.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>