sysvinit/inittab: adjust shutdown
The previous shutdown sequence was basically a no-op, it did't call rcK
to stop anything and dumped users back into shell.
So do the proper thing by calling rcK for levels 0 (halt) and 6
(reboot), then swapoff, then remount ro.
After that call halt for 0 and reboot for 6.

Also make 3 the default runlevel as in normal distributions, for all
intents and purposes it changes nothing except aesthetics/convention.

And make the labels a context sequence rather than just a short form
of the actions which IMHO looks horrible.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>