ci20_defconfig: remove BR2_ROOTFS_DEVICE_CREATION_DYNAMIC_MDEV=y
We had this option in the defconfig because the kernel defconfig we
provide doesn't have CONFIG_DEVTMPFS and CONFIG_DEVTMPFS_MOUNT, so a
/dev management which creates the devices dynamically using only
devtmpfs would not work. However, CONFIG_DEVTMPFS and
CONFIG_DEVTMPFS_MOUNT are automatically added to the kernel config when
you build the kernel using Buildroot, except when you choose
BR2_ROOTFS_DEVICE_CREATION_STATIC as a /dev management. This is handled
by linux/linux.mk.

Signed-off-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>