package/erlang-rebar: new host package
rebar is used to compile Erlang packages.

We need  host variant so as to be able to provide it to Erlang packages
that do not bundle their own version, or bundle a broken version.

Since this is a host-only package, used only internally, we do not
provide a Kconfig option for it. Packages that need it will depend on
it.

Signed-off-by: Johan Oudinet <johan.oudinet@gmail.com>
[yann.morin.1998@free.fr: extract host-rebar to its own patch]
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>