libxmlrpc: improve curl-config patch
The curl-config patch that makes sure to not use curl-config from the
PATH but the curl-config program specified at configure time was
forgetting a number of other places. So we change the strategy, and
make the configure.in script define a $CURL_CONFIG variable that can
be used in Makefiles where appropriate.

This means we need to run autoconf (but not AUTORECONF = YES since the
package uses autoconf but not automake).

This fixes build failures in cases where a curl-config program is in
the PATH. Fixes:

  http://autobuild.buildroot.org/results/06b/06b7eada721483eb57adde147768a1f306b7d0d7/
  (and many similar ones)

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Tested-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Acked-by: "Yann E. MORIN" <yann.morin.1998@free.fr>