package/infozip: fix LARGE_FILE_SUPPORT check
The LFS check was dropped in commit cc4dedd9.

Since then there is a runtime error:
zip warning : Not supported (uzoff_t not same size as zoff_t)

However LFS check can't readded as is because it can't work
for cross-compilation since it try to run a target's binary on
the host system.

Replace infozip's LFS check with the one used by autoconf.
This allows to readd LFS support for host-infozip.

Reported-by: Richard Genoud <richard.genoud@gmail.com>
Signed-off-by: Romain Naour <romain.naour@openwide.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>