package/vlc: bump to 2.2.0
Since we bumped FreeRDP, vlc no longer builds:
    http://autobuild.buildroot.org/results/8d9/8d91e3ffccee753135e9d0faa2b1ae00875606c9/
    http://autobuild.buildroot.org/results/e1f/e1f6526d1a4ac2c95998793efea2188faa708e29/

Fix that by bumping vlc to 2.2.0, as vlc-2.1.6will never support this
new version of FreeRDP.

Add dependency on kernel headers >= 3.7, because it needs DTV_STREAM_ID.
I stumbled on this whith a toolchain using headers forom 3.5, but only
tested with a toolchain with headers from 3.9, so there might be other
dependencies I missed on 3.8 or even 3.9. Let's wait for build failures
to find out...

There are quite a bunch of new --enable/disable options, so I've tried
to catter for all the changes. Most notably, there's no longer a flag to
enable.disable GL, it is now always auto-detected. Also, I've added
flags for GLES which we were missing so far, so we had to rely on the
build order to detect those dependencies (which was OK becasue libgles
sorts before vlc). Finally, SVG now has two flags to enable different
parts of SVG support, but they both depend on the same library, so
enable both.

There are new support for some features, like acceleration on RPI, but
that can well wait for further patches when someone is interested ;-) ,
vdpau HW accelearation on NVidia (we have a package for that, but not
able to test, so not added), vpx support (we have a libvpx package,
but I had no time to test), x262 and x265 support for which we're
currently lacking the packaging, and a few other exotic stuff...

Reported-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>