package/powertop: handle ncursesw properly and depend on wchar
The powertop code uses some wide-char functions, so the package should
depend on wide-char support. Also, when wide-char support is
available, the wide-char variant of ncurses may be built, but powertop
detection logic isn't quite up to the game, so we help him a little
bit.

Thanks to Gustavo Zacarias for doing some additional investigation and
proposing the solution implemented in this patch.

Fixes
http://autobuild.buildroot.net/results/913/913cea22f8a8f5902d8da5f64c3fce056d66790f/
http://autobuild.buildroot.net/results/6e1/6e11fa2a7405a69c59ced046b92ff08660c4aab7/
http://autobuild.buildroot.net/results/1d3/1d3323b2afaefa7989854dbccf92015731199e66/
http://autobuild.buildroot.net/results/1e3/1e31d412d8b3a38a375ad0be8f696bee993ec297/
http://autobuild.buildroot.net/results/2ff/2ff511eb8d00b94aca68427446e2d0f6e4317a5a/
and maybe others

[Thomas:
 - do not force select NCURSES_WCHAR, instead improve the detection
   logic to properly use it when available. Solution provided by
   Gustavo.
 - improve the commit log.]

Signed-off-by: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>