network-manager: needs (e)glibc
Fixes:
http://autobuild.buildroot.net/results/1cc/1ccd21cf2ae58ecac75a4cba393b7be4d2e43e5d/
http://autobuild.buildroot.net/results/718/718fc0eb28fc056ac83352547fe92313b33c3d59/
http://autobuild.buildroot.net/results/704/704bff70b73e77f552dc6b91732afefd846a1f50/
http://autobuild.buildroot.net/results/1ee/1eef17a793d151eb140d3364fbb8020e53eec10f/

And many more.

With the recent inclusion of systemd-dhcp, network-manager no longer builds
on uClibc/musl as it uses:

- CLOCK_BOOTTIME (not available in uClibc 0.9.33.2)
- IPTOS_CLASS_* (not available in uClibc/musl)
- Uses locale functionality (may not be available in uClibc)

Now, patches could be written to make these things optional, but systemd is
historically not known to be receptive of such kind of patches (and NM wants
to be able to sync with systemd with minimal effort, so is not interested in
diverging from upstream), and NM is big / has lots of dependencies so just
mark it (e)glibc-only like systemd.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>