postgresql: avoid AC_TRY_RUN when testing snprintf()
A few snprintf() compatibility tests in config/c-library.m4 use AC_TRY_RUN
that doesn't work for cross compilation. As a result configure.in decides to
build a bundled implementation of snprintf(). This implementation calls
isnan(), but -lm is missing. This breaks link with libpq.so. Preseed the
correct values to make the configure script happy, and fix rsyslog build.

Fixes:
http://autobuild.buildroot.net/results/e76/e769982e3131581b38698c109c9bc5215e3d5b11/
http://autobuild.buildroot.net/results/2b9/2b9d3edc8341675455bc3b74d1e84c8c0c24c521/
http://autobuild.buildroot.net/results/2a8/2a80357237ed7225823a008b5ff02bab01269814/

and many more.

Signed-off-by: Baruch Siach <baruch@tkos.co.il>
Tested-by: Gergely Imreh <imrehg@gmail.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>