package/opencv: fix opencv.pc generation
This change fixes build of packages depending on opencv and using
pkg-config to get the opencv LDFLAGS, which were wrongly ordered for
static build and included absolute path pointing to the host system
libraries instead of the sysroot ones.

This patch fixes build issues with gst1-plugins-bad and vlc.

This change introduce and new patch to opencv, which is already merged
upstream [1].

Fixes:
  http://autobuild.buildroot.net/results/c70/c70f793c7c3b380428d6e2e622a90805cffc4dd1/
  ... and many others

[1] https://github.com/Itseez/opencv/commit/eceada586bbf18fc267e437522ec4f1f23ddc656

Cc: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Samuel Martin <s.martin49@gmail.com>
Tested-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>