package/matchbox: make matchbox-lib a real package
Currently, the matchbox package is using weird, legacy constructs to
build its different parts.

Notably, it adds matchbox-lib to the list of packages to build, even
though it is not a real package since it does not have a Kconfig entry.

Fix that:
  - add a Kconfig entry for matchbox-lib
  - select it from the top-level matchbox package

Note that matchbox-common already depends on matchbox-lib.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>