package/matchbox: drop useless dependencies
xlib-libXft is only needed when building a 'standalone' matchbox, which
means not using matchbox-lib.

But we *are* building matchbox-lib, and we do not support the standalone
mode (and probably won't, as even upstream says it is ugly, by lack of
theming).

Similarly, xlib-libXext is only used for its 'xsync' extension, for
which support is entirely commented-out in matchbox.

So, drop the dependency on xlib-libXft, and make it explicit we're not
building either standalone modes; drop xlib-libXext.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Reviewed-by: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>