package/matchbox-panel: use ACPI for the battery applet
It is possible to use ACPI to monitor the battery with the battery
mini-applet. If ACPI is not explicitly enabled, then APM is used if
available; if ACPI is not explcitly enabled and APM is not available,
then the battery applet is not built.

However, APM is not really current nowadays, and possibly missing for a
bunch of architecture...

What is a pity is that there is no way to explicitly enable or disable
the applet; it only relies on auto-detection... :-(

So, forcibly use ACPI on platforms that have it, that is x86 (32- and
64-bit) and AArch64.

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Arnout Vandecappelle <arnout@mind.be>
Reviewed-by: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>