flite: use FLITE_STRIP_COMPONENTS and FLITE_SUBDIR
So far, to overcome the buggy way the flite tarball is made, we had to
override the extract commands in a rather ugly way.

The newly introduced <PKG>_STRIP_COMPONENTS, along with <PKG>_SUBDIR and a
little edit to the patches, allow to remove the custom FLITE_EXTRACT_CMDS,
slightly simplifying the flite.mk code.

Signed-off-by: Luca Ceresoli <luca@lucaceresoli.net>
Cc: Samuel Martin <s.martin49@gmail.com>
Cc: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>