midori: bump to version 0.5.9
Bump to version 0.5.9 which uses the new webkit2 API (gtk3) or
alternatively fallback to webkit1 compatibility API (gtk2).
The old webkit is completely unsupported now, it makes no sense because
of instability and security issues.

Even though there's a newer midori 0.5.10 release it was skipped on
purpose since it's very unstable.

Signed-off-by: Gustavo Zacarias <gustavo@zacarias.com.ar>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>