mono: depend on shared library support
The mono source code includes <dlfcn.h>, which is only available when
shared library support is available. While it might be possible to do
a fully static installation of Mono, it probably isn't very useful.

While we're at it, this commit also makes sure that the Config.in
comment is not visible when the architecture doesn't support Mono.

Fixes:

  http://autobuild.buildroot.net/results/5d99bdf77f1942fa403081267c362aa1f8fd0dab/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>