x264: unbreak x86 build after version bump
Fixes:
http://autobuild.buildroot.net/results/3ec/3ec54f722d6008fc422540d3a5462b306d16e84c/

The recent x264 version bump broke the configure step on x86/x86-64 as x264
ends up using gas instead of yasm as assembler.  The reason for this is the
recent upstream commit to optionally use nasm instead of yasm if AS= is
passed:

commit b568a256b9bc6c500d7b1ffe4b9c3311ee5ff337
Author: Henrik Gramner <henrik@gramner.com>
Date:   Sat May 23 19:44:16 2015 +0200

    x86: Experimental nasm support

    Enables the use of nasm as an alternative to yasm.

    Note that nasm cannot assemble x264 with PIC enabled since it currently doesn't
    support [symbol-$$] addressing which is used extensively by x264's PIC code.
    This includes all 64-bit Windows and 64-bit OS X builds, even non-shared.

    For the above reason nasm is currently intentionally not auto-detected, instead
    the assembler must be explicitly specified using "AS=nasm ./configure".

    Also drop -O2 from ASFLAGS since it's simply ignored anyway.

But as we pass AS=$(TARGET_AS) it ends up using gas instead. Fix it by
explicitly passing AS=yasm instead.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>