qt5base: add optional gstreamer1 support
Indirectly fixes:
http://autobuild.buildroot.net/results/773/773d9996bd3da283eed6b75083b8bab6469d0423/
http://autobuild.buildroot.net/results/069/069f564fde821b55e710fbe93fe30a5cf11a6da5/
http://autobuild.buildroot.net/results/456/456d380f8b4478245e89f711c33fe0c2d23aaa3a/
http://autobuild.buildroot.net/results/7c8/7c8ebfc93a5b11e40251495715d742520b3b8291/

And many more, as qt5multimedia needs qt5base to be built with gstreamer1
support as well.

While qt5 has some gstreamer 0.10 support, it doesn't seem to be used much
any more and E.G. the 0.10 fallback code in qt5multimedia is broken so only
enable it for the 1.x series.

Signed-off-by: Peter Korsgaard <peter@korsgaard.com>