fastd: add new package
This is a new package for the fastd "Fast and Secure Tunneling Daemon"
which was developed for the Freifunk Gluon project in the first place.

It includes a patch to allow cross compiling with toolchains without
LTO support which fails with the unpatched version due to some ugly
cmake hacks in fastd v17, details in the patch.

[Thomas:
 - Get rid of trailing spaces in Config.in
 - Remove the BR2_PACKAGE_FASTD_OPENSSL, and simply rely on
   BR2_PACKAGE_OPENSSL
 - Remove -DWITH_CAPABILITIES=TRUE, since libcap support is anyway
   mandatory.
 - Use ON/OFF instead of TRUE/FALSE.]

Signed-off-by: Alexander Dahl <post@lespocky.de>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>