ffmpeg: mark as not available on nios2
ffmpeg was already marked as not available for the NIOS2 Sourcery
toolchains, but it could still be built with the internal toolchain
backend or a custom external toolchain.

However, an inspection of the latest glibc source code indicates that
FE_INVALID, FE_OVERFLOW and FE_UNDERFLOW and indeed not available in
the nios2 variant of <fenv.h>.

Consequently, this patch makes ffmpeg not available on nios2, which
allows to simplify a bit the dependencies.

It propagates this dependency to:

 - minidlna (and at the same time makes sure the minidlna comment is
   not displayed on nios2, which wasn't properly taken into account
   until now)
 - mpd
 - opencv
 - opencv3
 - squeezelite
 - tovid

Even if it selects ffmpeg, Kodi does not need an update since Kodi is
only available on a limited number of architectures (which don't
include nios2, obviously). Other packages only make use of ffmpeg when
available.

Fixes:

  http://autobuild.buildroot.org/results/921/9212f5a6432c5e695ac0630695405cea05e28610/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>