mongrel2: add upstream patch to fix symbol conflict
The symbol 'max_align_t', defined in the mongrel2 sources
(src/mem/align.h), is also defined in one of the standard headers in
some toolchains (the ones based on GCC 5+ apparently).

This conflict has been solved upstream, and a new version (1.10.0)
containing the fix has been released already, so this backported patch
may be dropped in the future when mongrel2 gets a version bump.

Fixes:

  http://autobuild.buildroot.net/results/b7f/b7f86bdda094de96633e9921e0e5895eadc89217/
  http://autobuild.buildroot.net/results/f62/f623703aa6d7992b67d305331734e4ed0c7b2be4/
  http://autobuild.buildroot.net/results/eb3/eb385cb4f87078122028e85cbe31caa01d794991/
  ...

Signed-off-by: Rodrigo Rebello <rprebello@gmail.com>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>