package/unscd: new package
A daemon which handles passwd, group and host lookups for running
programs and caches the results for the next query. This package is
intended to replace glibc nscd. Presently, glibc nscd is not installed
by Buildroot.

unscd depends on glibc because it relies on glibc function
__nss_disable_nscd.

nscd.conf is taken from glibc with unused configurations removed.

Cc: Yann E. MORIN <yann.morin.1998@free.fr>
Cc: Arnout Vandecappelle <arnout@mind.be>
Cc: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Signed-off-by: Doug Kehn <rdkehn@yahoo.com>
Reviewed-by: Arnout Vandecappelle (Essensium/Mind) <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>