python-m2crypto: different fix for the OpenSSL problem
m2crypto provides a --openssl option to specify the location of the
OpenSSL library, when not installed in /usr. However, this option does
not apply to the standard "build" command normally called by the
python-package infrastructure, but only to the "build_ext" command.

This already required to override the default python-package build
commands for this package. Moreover, this the last bump to version
0.22.5, things got even more complicated since calling just
"build_ext" does not work, and additional quirks are needed.

This commit proposes to get away with this by relying on the default
python-package behavior, and simply hardcoding the OpenSSL location by
patching the setup.py script in a post-patch hook.

This issue has also been reported upstream at
https://gitlab.com/m2crypto/m2crypto/issues/89.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>
Reviewed-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>