dbus: audit support needs both audit and libcap-ng
As was suggested by the <pkg>_DEPENDENCIES variable, the audit support
in DBus requires both audit and libcap-ng. However, it didn't take
care of the fact that libcap-ng must be enabled in the configuration
to depend on it, causing some build failures with the newly added
check.

DBus configure.ac confirms that both packages are needed to enable
audit support, so we simply fix the condition to only be true when
both BR2_PACKAGE_AUDIT *and* BR2_PACKAGE_LIBCAP_NG are true.

Fixes:

  http://autobuild.buildroot.org/results/239/23953cc66faecb65e9ebf1f6980924f823d736a2/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>