package/owfs: fix link with FUSE
owfs reinvents its own weird and wicked mechanism to find libfuse. It
believes that, by default, FUSE is installed in /usr/local, and forcibly
uses that path to look for headers and libs.

This is obviously wrong in cross-compilation, and triggers our check for
unsafe header/library paths.

The best option to fix that would be to use pkg-config to find FUSE, but
this is a bit involved. Instead, use config options to specify the path
to the headers and libs.

Fixes:
    http://autobuild.buildroot.org/results/2c8/2c8baf6910d7ac2e3c92efa62cc3b9d3737e1fb5/
    http://autobuild.buildroot.org/results/1f9/1f994f53ceca434ebd1d87923f85b5ed71229d94/
    and a few others...

Signed-off-by: "Yann E. MORIN" <yann.morin.1998@free.fr>
Cc: Arnout Vandecappelle <arnout@mind.be>
Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>