cannelloni: needs dynamic library
cannelloni selects lksctp which depends on !BR2_STATIC_LIBS, but this
dependency was not propagated to cannelloni. Due to this, when
cannelloni was selected, lksctp-tools could be built even in a
BR2_STATIC_LIBS=y configuration.

We fix this by propagating the dependency. While we're at it, we
adjust the indentation of the cannelloni comment dependencies.

Fixes:

  http://autobuild.buildroot.org/results/b589af984499c532951217eb500ff31e6655c63f/

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>