valgrind: disable for MIPS soft-float
As stated here [1], recent changes on the MIPS binutils sources have
made it necessary for GCC to pass the -msoft-float to the assembler. Due
to that, valgrind fails to build for MIPS soft-float when using a
version of binutils >= 2.25 because its using some hard-float
instructions.

However, we cannot just disable it for the soft-float and binutils >=
2.25 combination since external toolchains don't provide information
about the binutils version they use. So, instead, we simply disable it
for soft-float.

1: https://gcc.gnu.org/ml/gcc-patches/2014-08/msg00905.html

Fixes:
  http://autobuild.buildroot.net/results/5f5/5f576c7f8d56058a19ed0e7ff4b1ec620bcafb65/

Signed-off-by: Vicente Olivert Riera <Vincent.Riera@imgtec.com>
Signed-off-by: Peter Korsgaard <peter@korsgaard.com>