oprofile: not available on Microblaze
Commit c45979c732cb610ad5d54e23dd7d4d49e519d45b marked OProfile as not
available on the Xtensa architecture, due to the lack of memory
barrier operations. This commit does the same for the Microblaze
architecture, for the same reason, which allows to fix the following
autobuilder failure:

  http://autobuild.buildroot.org/results/9a872ddc906e9d552d30762e849a1b537b4e5095/

It is worth mentioning that most likely Xtensa and Microblaze are
architectures implementing a strongly-ordered memory model, in which
case we could define the memory barriers as no-ops. But until someone
who actually cares about OProfile on Xtensa and Microblaze shows up,
it's probably better to disable the package on those architectures.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>